Move ServerConfig and EndpointConfig into shared module
